Separando las classes y widgets para que se vea mejor en el responsive y para performance
