+++ lib/main.dart
@@ -1,146 +1,10 @@
-import 'package:cfmkids/splashscreen.dart';
-import 'package:cfmkids/login.dart';
-import 'package:cfmkids/signup.dart';
 import 'package:flutter/material.dart';
+
+import 'screens/homePage.dart';
 
 void main() {
   runApp(const MaterialApp(
     debugShowCheckedModeBanner: false,
     home: HomePage(),
   ));
-}
-
-class HomePage extends StatelessWidget {
-  const HomePage({super.key});
-
-  // This widget is the root of your application.
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Flutter Demo',
-      theme: ThemeData(
-        // This is the theme of your application.
-        //
-        // Try running your application with "flutter run". You'll see the
-        // application has a blue toolbar. Then, without quitting the app, try
-        // changing the primarySwatch below to Colors.green and then invoke
-        // "hot reload" (press "r" in the console where you ran "flutter run",
-        // or simply save your changes to "hot reload" in a Flutter IDE).
-        // Notice that the counter didn't reset back to zero; the application
-        // is not restarted.
-        primarySwatch: Colors.blue,
-      ),
-      home: SplashScreen(),
-    );
-  }
-}
-
-class MyHomePage extends StatefulWidget {
-  const MyHomePage({super.key, required this.title});
-
-  // This widget is the home page of your application. It is stateful, meaning
-  // that it has a State object (defined below) that contains fields that affect
-  // how it looks.
-
-  // This class is the configuration for the state. It holds the values (in this
-  // case the title) provided by the parent (in this case the App widget) and
-  // used by the build method of the State. Fields in a Widget subclass are
-  // always marked "final".
-
-  final String title;
-
-  @override
-  State<MyHomePage> createState() => _MyHomePageState();
-}
-
-class _MyHomePageState extends State<MyHomePage> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-        body: SafeArea(
-            child: Container(
-                decoration: const BoxDecoration(
-                    image: DecorationImage(
-                        image: AssetImage('assets/bg1920x1080.jpg'),
-                        fit: BoxFit.cover)),
-                width: double.infinity,
-                height: MediaQuery.of(context).size.height,
-                padding:
-                    const EdgeInsets.symmetric(horizontal: 30, vertical: 50),
-                child: Column(
-                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                  crossAxisAlignment: CrossAxisAlignment.center,
-                  children: <Widget>[
-                    Container(
-                      height: MediaQuery.of(context).size.height / 3,
-                      decoration: const BoxDecoration(
-                          color: Color.fromARGB(0, 185, 34, 34),
-                          image: DecorationImage(
-                              image: AssetImage('assets/logo-v1.png'))),
-                    ),
-                    Column(
-                      children: <Widget>[
-                        MaterialButton(
-                            key: const Key('loginBtn'),
-                            minWidth: double.infinity,
-                            height: 60,
-                            onPressed: () {
-                              Navigator.push(
-                                  context,
-                                  MaterialPageRoute(
-                                      builder: ((context) =>
-                                          const LoginPage())));
-                              ScaffoldMessenger.of(context).showSnackBar(
-                                const SnackBar(
-                                  content: Text("Login Button Pressed"),
-                                  duration: Duration(milliseconds: 300),
-                                ),
-                              );
-                            },
-                            shape: RoundedRectangleBorder(
-                                borderRadius: BorderRadius.circular(50),
-                                side: const BorderSide(color: Colors.black)),
-                            child: const Text("Login",
-                                style: TextStyle(
-                                    fontWeight: FontWeight.w600,
-                                    fontSize: 18))),
-                        const SizedBox(
-                          height: 20,
-                        ),
-                        Container(
-                            padding: const EdgeInsets.only(top: 3, left: 3),
-                            decoration: BoxDecoration(
-                              borderRadius: BorderRadius.circular(50),
-                            ),
-                            child: MaterialButton(
-                                key: const Key('signupBtn'),
-                                minWidth: double.infinity,
-                                height: 60,
-                                onPressed: () {
-                                  Navigator.push(
-                                      context,
-                                      MaterialPageRoute(
-                                          builder: ((context) =>
-                                              const SignupPage())));
-                                  ScaffoldMessenger.of(context).showSnackBar(
-                                    const SnackBar(
-                                      content: Text("Sign Up Button Pressed"),
-                                      duration: Duration(milliseconds: 300),
-                                    ),
-                                  );
-                                },
-                                color: Colors.blueGrey,
-                                elevation: 0,
-                                shape: RoundedRectangleBorder(
-                                  borderRadius: BorderRadius.circular(50),
-                                ),
-                                child: const Text("Sign up",
-                                    style: TextStyle(
-                                        fontWeight: FontWeight.w600,
-                                        fontSize: 18))))
-                      ],
-                    )
-                  ],
-                ))));
-  }
 }
