+++ lib/screens/signup.dart
@@ -1,0 +1,31 @@
+import 'package:flutter/material.dart';
+
+import '../widgets/signUpW.dart';
+
+class SignupPage extends StatelessWidget {
+  const SignupPage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+        resizeToAvoidBottomInset: false,
+        backgroundColor: Colors.transparent,
+        // fix the appbar button exhibition - it has to be on top of the container below it
+        // fix the visualization - keyboard covering 'Confirm Password' field
+        appBar: AppBar(
+          elevation: 0,
+          backgroundColor: Colors.white,
+          leading: IconButton(
+            onPressed: () {
+              Navigator.pop(context);
+            },
+            icon: const Icon(
+              Icons.arrow_back_ios,
+              size: 20,
+              color: Colors.black,
+            ),
+          ),
+        ),
+        body: const SignupWid());
+  }
+}
