+++ lib/screens/splashscreen.dart
@@ -1,8 +1,10 @@
 import 'package:flutter/material.dart';
 
-import 'responsive_login.dart';
+import '../responsive_login.dart';
 
 class SplashScreen extends StatefulWidget {
+  const SplashScreen({super.key});
+
   @override
   _SplashScreenState createState() => _SplashScreenState();
 }
