+++ test/widget_test_login.dart
@@ -1,14 +1,14 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:cfmkids/login.dart';
-import 'package:cfmkids/mainpage.dart';
+import 'package:cfmkids/screens/login.dart';
+import 'package:cfmkids/screens/mainpage.dart';
 
-// TODO: Add test cases for Sign up button
-// TODO: Add test cases to see if images displayed correctly
-// TODO: Add test cases to see if the layout is centered and correctly displaying padding and spacing
-// TODO: Add test cases to verify colors and texts for buttons
-// TODO: Add test cases to verify the app against different screen sizes and orientations
+// Add test cases for Sign up button
+// Add test cases to see if images displayed correctly
+// Add test cases to see if the layout is centered and correctly displaying padding and spacing
+// Add test cases to verify colors and texts for buttons
+// Add test cases to verify the app against different screen sizes and orientations
 
 void main() {
   testWidgets('Login button triggers the correct action',
@@ -37,9 +37,9 @@ void main() {
   testWidgets('Back button navigates back to the previous screen',
       (WidgetTester tester) async {
     // Build the LoginPage widget
-    await tester.pumpWidget(MaterialApp(
+    await tester.pumpWidget(const MaterialApp(
       home: Scaffold(
-        body: const Center(),
+        body: Center(),
       ),
     ));
